fix: changed data type for Project from int to Long.

+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/domain/model/commands/CreateMachineCommand.java
@@ -1,5 +1,5 @@
 package com.construtech.buildsphere.platform.resourceManagement.domain.model.commands;
 
 public record CreateMachineCommand(String machineName, String description, String receptionDate, String endDate,
-                                   double totalCost, int project) {
+                                   double totalCost, Long project) {
 }
