+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/domain/model/commands/CreateMaterialCommand.java
@@ -1,5 +1,5 @@
 package com.construtech.buildsphere.platform.resourceManagement.domain.model.commands;
 
 public record CreateMaterialCommand(String materialName, String description, String receptionDate,
-                                    int amount, double totalCost, String status, int project) {
+                                    int amount, double totalCost, String status, Long project) {
 }
